+++ src/Aardvark.Base.Essentials/Camera/CameraProjectionPerspective.cs
@@ -83,7 +83,7 @@ namespace Aardvark.Base
             get { return m_box.XY; }
             set
             {
-                if (m_box.IsInvalid) throw new ArgumentException();
+                if (value.IsInvalid) throw new ArgumentException();
                 m_box.Min.X = value.Min.X;
                 m_box.Min.Y = value.Min.Y;
                 m_box.Max.X = value.Max.X;